Updated to version 10.8.2 (code 324)
* In activealarm/NacActiveAlarmService.java:

    - Added suppress deprecation annotation as the deprecated method,
      stopForeground() was already being handled, but the build process
      was printing warnings as if it was not.

    - Fixed typo in a comment.

* In alarm/NacAlarmRepository.java:

    - Changed "mAllAlarms", the member variable, to final.

* In alarm/NacAlarmViewModel.java:

    - Changed "mAllAlarms", the member variable, to final.

* In file/browser/NacFileBrowser.java:

    - Fixed issue with calls to show() occasionally causing ANRs by
      wrapping:

      clearEntries(), populateEntries(), and getTree().cd(dir) in a
      Handler so that it would be at the back of the queue of the main
      thread.

* In main/NacMainActivity.java:

    - Renamed setupRefreshMainActivity() to refreshMainActivity() and
      removed the check before the activity is refreshed. Now, as the
      name suggests, the method will refresh the main activity so any
      checks need to happen before this method is called.

    - Removed check in setupForAppFirstRun. Now, any checks need to
      happen before this method is called.

    - In onChanged():

      Added check before calling setupForAppFirstRun().

      Added check before calling refreshMainActivity().

      Added more comments.

      Fixed formatting of javadoc.

    - Fixed issue where if the app was updated, then the alarm went off,
      then the user hit snooze, then the What's New dialog showed up,
      and the user interacted with it, the app would clear the snooze
      from the alarm. This is because in onReadWhatsNew(), there was a
      call to NacScheduler.refreshAll(). This is removed now.

    - In onResume():

      Moved call to refreshMainActivity() first so that if it needs to
      be called, nothing the method will just return and nothing else
      will be called.

      Added check before calling refreshMainActivity().

      Grouped like methods together such as the broadcast receiver
      setups and reorganized others.

      Added more comments.

    - Commented out printing of stack traces when the BroadcastReceivers
      would raise an IllegalArgumentException. That was for debugging
      before, but now it is not needed.

    - Removed unused imports.

* In media/NacMedia.java:

    - Added space between "if" and "(" in a couple places.

    - Removed extra newlines between imports.

* In media/NacMediaPlayer.java:

    - Changed "mMediaPlayer", the member variable, to final.

    - Added space between "if" and "(" in a couple places.

* In mediapicker/music/NacMusicFragment.java:

    - Removed extra newlines between imports.

* In name/NacNameDialog.java:

    - In onShowDialog():

      Wrapped showKeyboard() in a Handler and delayed it by 250 ms so
      that there is enough time for the TextInputEditText to have focus
      as well as the dialog window to have shown up and also gotten
      focus. These are the prerequisites before you can have a View
      request to show the keyboard.

      Added more comments.

* In nfc/NacNfc.java:

    - Added suppress deprecation annotation as the deprecated method,
      Intent.getParcelableExtra() was already being handled, but the
      build process was printing warnings as if it was not.

    - Removed extra newlines between imports.

* In scheduler/NacScheduler.java:

    - Removed cast of alarm ID to an int as it was unneeded.

* In system/NacBundle.java:

    - Added suppress deprecation annotation as the deprecated method,
      Bundle.getParcelable() was already being handled, but the build
      process was printing warnings as if it was not.

* In tts/NacTextToSpeech.java:

    - Added space between "if" and "(".

* In util/dialog/NacDialog.java:

    - In closeKeyboard() and showKeyboard(), added the new methods for
      acting on the keyboard since it was changed starting in
      Build.VERSION_CODES.S and above.

      Also added a suppress deprecation annotation as the build process
      was still printing warnings when it should not have been.

    - Removed extra newlines between imports.

* In values-b+es/strings.xml:

    - Added strings for the What's New dialog.

* In values/strings.xml:

    - Added strings for the What's New dialog.

+++ app/src/main/java/com/nfcalarmclock/scheduler/NacScheduler.java
@@ -137,7 +137,7 @@ public class NacScheduler
 		Intent intent = new Intent(context, NacMainActivity.class);
 
 		// Build the pending intent
-		return PendingIntent.getActivity(context, (int)id, intent , flags);
+		return PendingIntent.getActivity(context, id, intent , flags);
 	}
 
 	/**
